refactor: reorganize CLI structure and enhance platform detection
- Update CLI file structure to improve organization and maintainability.
- Introduce new platform detection logic to streamline binary path resolution.
- Remove deprecated fetch-cache module and integrate ambient type definitions.
- Enhance API route handling with new types and response utilities.
- Update tests to reflect changes in module imports and structure.
