--- packages/rari/src/fetch-cache.ts
@@ -1,11 +1,0 @@
-declare global {
-  interface RequestInit {
-    rari?: {
-      revalidate?: number | false
-      tags?: string[]
-      timeout?: number
-    }
-  }
-}
-
-export const rariFetchCacheTypes = true
