+++ packages/rari/vite.config.ts
@@ -7,8 +7,8 @@ export default defineConfig({
       'client': 'src/client.ts',
       'router': 'src/router.ts',
       'vite': 'src/vite.ts',
-      'cli': 'src/cli.ts',
-      'platform': 'src/platform.ts',
+      'cli': 'src/cli/index.ts',
+      'platform': 'src/cli/platform.ts',
       'image/index': 'src/image/index.ts',
       'og/index': 'src/og/index.ts',
       'mdx': 'src/mdx.ts',
